working on users management

+++ src/VC.Admin/Models/Infrastructure/FullReferenceDataModel.cs
@@ -8,6 +8,6 @@ namespace VitalChoice.Models.Infrastructure
     public class FullReferenceDataModel : RestrictedReferenceData
     {
 		public IList<LookupItem<string>> Roles { get; set; }
-	    public IList<LookupItem<int>> UserStatuses { get; set; }
+	    public IList<LookupItem<byte>> UserStatuses { get; set; }
     }
 }
